# contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Pedidos360 BFF
  version: 0.2.0
  description: API protegida con access tokens v2 de Microsoft Entra. Scopes delegados y rol Admin en la consulta administrativa. Configuración pública en config/entra-public.json.
servers:
  - url: https://7k2zyh0t0d.execute-api.us-east-1.amazonaws.com/api/v1
    description: Laboratorio AWS previamente verificado; requiere laboratorio activo
  - url: http://localhost:8180/api/v1
    description: Desarrollo local en WSL
security:
  - bearerAuth: []
paths:
  /me:
    get:
      operationId: getMyIdentity
      summary: Consultar identidad y permisos verificados por el BFF
      description: Requiere al menos uno de los tres scopes de Pedidos360. Respuesta Cache-Control no-store; nunca devuelve el token.
      responses:
        '200':
          description: Claims permitidos después de validar el JWT
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Identity' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /admin/pedidos:
    get:
      operationId: listTenantOrders
      summary: Consulta administrativa de pedidos del tenant autenticado
      description: Requiere Orders.Read Y rol Admin, comprobados por BFF y Orders. No modifica el filtro de propietario de /pedidos.
      responses:
        '200':
          description: Pedidos del tenant; misma representación de pedido, sin datos personales adicionales
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Order' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '502': { $ref: '#/components/responses/Unavailable' }
  /productos:
    get:
      operationId: listProducts
      summary: Listar productos del catálogo
      description: Requiere Catalog.Read.
      responses:
        '200':
          description: Catálogo
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Product' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '502': { $ref: '#/components/responses/Unavailable' }
  /productos/{id}:
    parameters:
      - $ref: '#/components/parameters/Id'
    get:
      operationId: getProduct
      summary: Consultar producto
      description: Requiere Catalog.Read.
      responses:
        '200':
          description: Producto
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Product' }
        '400': { $ref: '#/components/responses/InvalidRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '502': { $ref: '#/components/responses/Unavailable' }
  /pedidos:
    get:
      operationId: listMyOrders
      summary: Listar pedidos del propietario autenticado
      description: Requiere Orders.Read. No acepta un owner enviado por el cliente.
      responses:
        '200':
          description: Pedidos del propietario
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Order' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '502': { $ref: '#/components/responses/Unavailable' }
    post:
      operationId: createOrder
      summary: Registrar una solicitud multítem
      description: Requiere Orders.Create y Catalog.Read. Conserva snapshots de precio. No reserva stock ni admite reintento idempotente.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateOrder' }
            example:
              items:
                - productoId: 10000000-0000-0000-0000-000000000001
                  cantidad: 2
                - productoId: 10000000-0000-0000-0000-000000000002
                  cantidad: 1
      responses:
        '201':
          description: Solicitud registrada
          headers:
            Location:
              schema: { type: string }
              description: Ruta relativa /api/v1/pedidos/{id}
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Order' }
        '400': { $ref: '#/components/responses/InvalidRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '502': { $ref: '#/components/responses/Unavailable' }
  /pedidos/{id}:
    parameters:
      - $ref: '#/components/parameters/Id'
    get:
      operationId: getMyOrder
      summary: Consultar pedido propio
      description: Requiere Orders.Read. Un pedido de otro propietario responde 404.
      responses:
        '200':
          description: Pedido con snapshots persistidos
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Order' }
        '400': { $ref: '#/components/responses/InvalidRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '502': { $ref: '#/components/responses/Unavailable' }
components:
  securitySchemes:
    bearerAuth: { type: http, scheme: bearer, bearerFormat: JWT }
  parameters:
    Id:
      name: id
      in: path
      required: true
      schema: { type: string, format: uuid }
  responses:
    Unauthorized:
      description: Bearer ausente o JWT inválido; el cuerpo puede estar vacío.
    Forbidden:
      description: JWT válido sin los scopes o rol requeridos; el cuerpo puede estar vacío.
    NotFound:
      description: Recurso inexistente o pedido no perteneciente al usuario.
    InvalidRequest:
      description: UUID, cuerpo o regla de dominio inválidos.
    Unavailable:
      description: No fue posible completar la comunicación con un servicio interno.
  schemas:
    Identity:
      type: object
      required: [name, username, tenantId, subjectId, tokenSubject, issuer, audience, clientId, tokenVersion, roles, scopes, issuedAt, expiresAt]
      properties:
        name: { type: [string, 'null'], description: Solo presentación; no se usa para autorizar. }
        username: { type: [string, 'null'], description: Claim opcional de presentación. }
        tenantId: { type: string }
        subjectId: { type: string, description: oid usado junto al tenant para propiedad de pedidos. }
        tokenSubject: { type: [string, 'null'], description: sub del JWT; identidad pairwise de la API. }
        issuer: { type: string, format: uri }
        audience: { type: array, items: { type: string } }
        clientId: { type: string }
        tokenVersion: { type: string, enum: ['2.0'] }
        roles: { type: array, items: { type: string } }
        scopes: { type: array, items: { type: string } }
        issuedAt: { type: [string, 'null'], format: date-time }
        expiresAt: { type: string, format: date-time }
    Product:
      type: object
      required: [id, nombre, descripcion, precio, moneda, stock, disponible]
      properties:
        id: { type: string, format: uuid }
        nombre: { type: string, maxLength: 120 }
        descripcion: { type: string, maxLength: 500 }
        precio: { type: number, minimum: 0, multipleOf: 0.01 }
        moneda: { type: string, pattern: '^[A-Z]{3}$' }
        stock: { type: integer, minimum: 0, description: 'Existencia referencial, sin reserva.' }
        disponible: { type: boolean }
    CreateOrder:
      type: object
      required: [items]
      properties:
        items:
          type: array
          minItems: 1
          maxItems: 50
          description: No repetir productoId. Las líneas deben ser no nulas.
          items:
            type: object
            required: [productoId, cantidad]
            properties:
              productoId: { type: string, format: uuid }
              cantidad: { type: integer, minimum: 1, maximum: 1000 }
    OrderItem:
      type: object
      required: [productoId, nombre, precioUnitario, cantidad, subtotal]
      properties:
        productoId: { type: string, format: uuid }
        nombre: { type: string, description: Snapshot al registrar. }
        precioUnitario: { type: number, minimum: 0, multipleOf: 0.01 }
        cantidad: { type: integer, minimum: 1, maximum: 1000 }
        subtotal: { type: number, minimum: 0, multipleOf: 0.01 }
    Order:
      type: object
      required: [id, estado, fechaCreacion, items, total, moneda]
      properties:
        id: { type: string, format: uuid }
        estado: { type: string, enum: [REGISTERED] }
        fechaCreacion: { type: string, format: date-time, description: 'UTC, precisión máxima de microsegundos.' }
        items:
          type: array
          minItems: 1
          maxItems: 50
          items: { $ref: '#/components/schemas/OrderItem' }
        total: { type: number, minimum: 0, multipleOf: 0.01 }
        moneda: { type: string, pattern: '^[A-Z]{3}$' }